requirements.txt
streamlit
langchain
faiss-cpu
openai
sentence-transformers
numpy